O.K. Fran if I messed anything up then I apologise. I changed the initialise db script so that it creates the Habit table . Cos otherwise it errored for me. There may be something here I don't understand properly.
I added knockout and a really simple in page updating list.

Plan is to hook this up to the db and then decide on a testing strategy.

Overall I would go with webdriver for some ui whole stack tests. And some jasmine for the javascript.

The new route is called summary.
It has it's own template.

Feel free to edjumacate me about the pythons.